Do not use converter on BsonValue types.
Referenced assembly was not loaded in a reflection-only context, but
then chose to load the same assembly again in a reflection-only
context and do conversions between them. This led to a strange error
message of received type 'MongoDB.Bson.BsonValue' but expected
'MongoDB.Bson.BsonValue' in the Expr.Call section.